[add] 1.新增Data Server 2.新增redis数据缓存 3.新增MySql数据持久化 4.新增Client---GATE---GS消息转发 5.新增UUID生成函数 [fix] 1.修复gate建立连接FD未设置非阻塞 2.修复evbuffer读完后未设置读完标志 [mod]

+++ src/common/utils/str_util.cpp
@@ -18,7 +18,7 @@ const std::string& BlankStr() {
     return s_blank;
 }
 
-std::string int_to_str(int val) {
+std::string int_to_str(int64_t val) {
     std::stringstream ss;
     ss << val;
     return ss.str();
