+++ src/common/utils/str_util.h
@@ -2,6 +2,7 @@
 #define STR_UTIL_H
 
 #include <string>
+#include <stdint.h>
 
 namespace util {
 namespace str {
@@ -10,7 +11,7 @@ const char* Blank();
 
 const std::string& BlankStr();
 
-std::string int_to_str(int val);
+std::string int_to_str(int64_t val);
 
 int str_to_int(const char* str);
